Introduce binary execution command
